+++ orders/models.py
@@ -96,8 +96,8 @@ class OrderItem(models.Model):
     weight = models.FloatField(
         'Вес', default=0, validators=[MinValueValidator(0)]
     )
-    size = models.IntegerField(
-        'Размер', default=0, validators=[MinValueValidator(0)]
+    size = models.FloatField(
+        'Размер', default=0.0, validators=[MinValueValidator(0)]
     )
     quantity = models.PositiveIntegerField('Количество', default=1)
     price = models.DecimalField(
